feat: migrate CollisionSystem.test.ts to pure business logic layer
- Extract pure business logic functions to src/utils/CollisionUtils.ts
- Create src/test/pure/business/CollisionSystem.test.ts with 33 tests
- Update original CollisionSystem.test.ts to keep only 13 framework-dependent tests
- Maintain sub-second execution time (853ms) and 99 total pure logic tests
- Follow sextuple-proven migration pattern for Phase 2 continuation
